Progress Update
Implemented the Javascript component of the conncurrent execution feature for the Visual Programming interface.
Python code translation is in progressed.

Three new blocks are added in order to support this feature:

1. dispatch operation block: This block takes in regular programming and A.I. blocks as its children blocks.
And execute these blocks in parallel with the current main program. In Javascript, this is implemented by using
JS-interpreter. In Python, this is implemented by using threads.

2. set main process variable block: This block can be used within the dispatch operation block to change the
values of the variables in the main program. This is done by adding the new values to a variable update queue
in the main program. The main program can use this queue and update the coressponding variable values using the
processing loop block.

3. processing loop block: This block acts as a while loop. But in the end of the loop, it scans any change in
the variable update queue, if there are updates available, it will update the corresponding variables with the
new values.