Support the use of global function in dispatch block in visual programming interface
Before this commit, only variable from the main program can be used within a dispatch block. This commit allows users
to use functions that are defined in the main program within a dispatch block. This is done by passing in the full
definition of all user defined functions as code string to the dispatch thread.