Fixed setup screen crashed issue
When the setup page tried to read the hostname from the current browser, it assumed the hostname has a ".local" in it. However, this is not true, when the hostname returned from the broswer does not contain ".local", the setup screen crashed.
This fix removed such assumption and handled all kinds of valid hostnames.