Add test scripts and fix Javascript syntax bug
A set of test scripts are added to the test directory.

When translating multiple dispatch blocks into Javascript code, there is a bug causing the translated code to include an extra " symbol
in subsequent dispatch blocks. This caused syntax error when the users start running the program. This is fixed in this commit.