Fixed a bug that occurs when user uses IP address to open CAIT in a browser.
When user uses the Pi's IP address to login CAIT, the page will be stuck with a "Starting CURT Backend" backend message.
This is due to a mismatch between the IP address and CAIT's internal device hostname. CAIT's backend uses the device's hostname to
broadcast the status updates to the browser frontend, when user uses IP address to login CAIT in the browser, the frontend could not match this IP address to the
device's hostname, and thus couldn't receive any status update from CAIT. This is fixed by making the CAIT backend to
broadcast one more message to let the frontend to know how to match the IP address with a device hostname.