Add exception handling code for FileExistsError when loading a TVM model with Docker Container.
Since the model directory of CEP is mounted to the vision Docker Container, there is a chance that when
loading a model using TVM in the vision container, the temporary directory created during the loading
process may not be properly deleted. This is because the vision container maybe pointing to an older
inode. If this happens, the directory delete operation at the end of model loading will fail.
And since TVM uses the ignore_error=True flag during the delete process, this failure will not be caught.
When this scenario happens, TVM will not be able to load the model in the next restart and throws a
FileExistsError as it cannot create the temporary directory. This commit implemented the handling of
the FileExistsError. It first removes the existing temporary directory and then loads the model again.
The handling process will retry loading the model up to three times. If the loading still faile,
it will throw an exception.